refactor: 💡 nicer type guards
